[ksmart02-KKM]멘토 경력 등록 페이지 detail: 멘토 수정, 삭제 페이지 기능 해야함

+++ ks52team02/src/main/java/ks52team02/member/mypage/dto/MentorEducation.java
@@ -8,7 +8,7 @@ public class MentorEducation {
 	private String mentorEducationCode;
 	private String mentorId;
 	private String mentorAcbgUp;		//회종학력유무
-	private int mentorCredits;			//학점
+	private float mentorCredits;		//학점
 	private String mentorSchoolNm;		//학교명
 	private String mentorScsbjtNm;		//전공
 	private String mentorSection;		// 졸업/수료 구분
